Much more work on mappers.

+++ src/core/parser.rs
@@ -1,7 +1,7 @@
 use core::cpu::*;
 use core::ines::INES;
 
-pub fn parser(address: usize, ines: &INES) -> AsmInstruction {
+pub fn parser(address: u16, ines: &INES) -> AsmInstruction {
     let opcode = ines.read(address);
     AsmInstruction::new(opcode)
 }
